requirements.txt
Django==1.5.5
South
python-dateutil
django-extensions
ipython
markdown
chardet
git+git://github.com/twidi/ghdiff.git#egg=ghdiff
django-jsonfield
git+ssh://git@github.com/twidi/pytimeago.git#egg=pytimeago
django-debug-toolbar
git+git://github.com/twidi/django-debug-toolbar-template-timings.git#egg=template_timings
git+git://github.com/twidi/django-decorator-include.git#egg=django-decorator-include
git+git://github.com/twidi/django-extended-choices.git#egg=django-extended-choices
-e git+git://github.com/twidi/django-template-preprocessor.git#egg=django-template-preprocessor
hiredis
redis-limpyd-jobs
django-async-messages
django-redis
django-adv-cache-tag
whatthepatch